+++ Cargo.toml
@@ -11,6 +11,7 @@ members = [
 	"demo",
 	"ljx",
 	"liblogjet",
+	"ljcel",
 	"plugins/syslog-ingest",
 	"plugins/logcat-ingest",
 	"plugins/stress-ingest",
@@ -28,6 +29,7 @@ default-members = [
 	"demo",
 	"ljx",
 	"liblogjet",
+	"ljcel",
 	"plugins/syslog-ingest",
 	"plugins/logcat-ingest",
 	"plugins/stress-ingest",
